backend/requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.36.0
python-dotenv==1.1.1
supabase==2.22.0
